moved serialization code into separate file
